Add turnorder buttons to theme
Start turn and actions buttons now use primary and secondary colors and are all the same color. The enemy and legendary checks are still in there if custom action colors for those types of tokens want to be added to the theme.

Hovering over Active Turn changes text to End Turn.